requirements.txt
Django==3.1
matplotlib
django-bootstrap4
django-bootstrap-datepicker-plus
